# backend/src/main/java/com/pas/backend/dto/user/UserCreateRequest.java
package com.pas.backend.dto.user;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class UserCreateRequest {

    @NotBlank(message = "Employee code is required")
    private String employeeCode;

    @NotBlank(message = "Full name is required")
    private String fullName;

    @NotBlank(message = "Email is required")
    @Email(message = "Email must be valid")
    private String email;

    @NotBlank(message = "Password is required")
    private String password;

    @NotBlank(message = "Role is required")
    private String role;

    private String phoneNumber;

    private String department;

    private java.util.UUID positionId;

    private Float initialExperienceYears;

    private Boolean isActive;
}