# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hệ thống Chấm công QR API
  description: |
    API for QR Code Attendance System with GPS verification.

    ## Authentication
    All endpoints require JWT Bearer token except `/auth/login`.

    ## Rate Limiting
    - Authentication: 5 requests/minute
    - Attendance: 10 requests/minute
    - General: 100 requests/minute
  version: 1.0.0
  contact:
    name: API Support
    email: support@company.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development
  - url: https://attendance.company.com/api/v1
    description: Production
...
paths:
...
  /users:
    get:
      tags: [Users]
      operationId: getUsers
      summary: Get all users
      description: Get list of all users (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: role
          in: query
          schema:
            type: string
            enum: [EMPLOYEE, ADMIN, HR_MANAGER]
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, LOCKED]
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      operationId: createUser
      summary: Create new user
      description: Create new user account (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Get user by ID
      description: Get user details (Admin or self)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Users]
      operationId: updateUser
      summary: Update user
      description: Update user information (Admin or self)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete user
      description: Delete user account (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: User deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /shifts:
    get:
      tags: [Config]
      operationId: getShifts
      summary: Get all shifts
      description: Get list of work shifts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Shift list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftResponse'

    post:
      tags: [Config]
      operationId: createShift
      summary: Create new shift
      description: Create new work shift (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftRequest'
      responses:
        '201':
          description: Shift created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  schemas:
...
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        shiftId:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    UserCreateRequest:
      type: object
      required: [username, password, fullName, role]
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [EMPLOYEE, ADMIN, HR_MANAGER]
        shiftId:
          type: integer
    UserUpdateRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        shiftId:
          type: integer
        status:
          type: string
    UserPageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/User'
            page:
              type: integer
            size:
              type: integer
            totalElements:
              type: integer
            totalPages:
              type: integer
...
    ShiftResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        gracePeriod:
          type: integer
        workDays:
          type: string
        isActive:
          type: boolean
    ShiftRequest:
      type: object
      required: [name, startTime, endTime, workDays]
      properties:
        name:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        gracePeriod:
          type: integer
          default: 15
        workDays:
          type: string
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...